src/auth-service/requirements.txt
astroid
cffi
click
cryptography
dill
Flask
importlib-metadata
isort
itsdangerous
jedi
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
parso
platformdirs
psycopg2-binary
pycparser
PyJWT
pylint
tomli
tomlkit
typing-extensions
Werkzeug
wrapt
zipp
flask-cors
python-dotenv
bcrypt
prometheus-client
pytest
black
flask_prometheus_metrics
python-json-logger
